+++ presentation/src/main/java/com/anytypeio/anytype/presentation/page/PageViewModel.kt
@@ -581,7 +581,13 @@ class PageViewModel(
         if (editorHasChildrenScreens) {
             dispatch(Command.PopBackStack)
         } else {
-            proceedWithExiting()
+            val state = controlPanelViewState.value
+            checkNotNull(state) { "Control panel state is null" }
+            if (state.stylingToolbar.isVisible) {
+                onCloseBlockStyleToolbarClicked()
+            } else {
+                proceedWithExiting()
+            }
         }
     }