+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Images can be full screen when tapped, allowing users to zoom in (#968)
 * Disable animation for edit-mode in order to increase editor performance (#884)
+* Should close style toolbar (instead of closing document) when back button pressed (#973)
 
 ### Design & UX 🔳